Improved the way in which newline characters are escaped.
The "VCardStringUtils.escapeText()" method no longer escapes newlines.  Instead, newlines are escaped within the "VCardWriter" class when the vCard is marshaled.  This reduces errors because the type classes are no longer responsible for escaping newlines (so, an extended type class that forgets to call "escapeText()" on its value within the "doMarshalValue()" method will no longer break the entire vCard).